refactor composeObject to ts
