+++ examples/compose-object.mjs
@@ -51,7 +51,7 @@ const sourceList = [
 
 const destOption = new Minio.CopyDestinationOptions({
   Bucket: bucketName,
-  Object: '100MB.zip',
+  Object: 'object-name',
   /** Other possible options */
   /* Encryption:{
         type:Helpers.ENCRYPTION_TYPES.KMS,
@@ -67,11 +67,4 @@ const destOption = new Minio.CopyDestinationOptions({
   */
 })
 
-const composePromise = s3Client.composeObject(destOption, sourceList)
-composePromise
-  .then((result) => {
-    console.log('ComposeObject Success...', result)
-  })
-  .catch((e) => {
-    console.log('composeObject Promise Error', e)
-  })
+await s3Client.composeObject(destOption, sourceList)
