+++ src/minio.js
@@ -16,18 +16,14 @@
 
 import * as Stream from 'node:stream'
 
-import async from 'async'
-import _ from 'lodash'
 import * as querystring from 'query-string'
 import xml2js from 'xml2js'
 
 import * as errors from './errors.ts'
-import { CopyDestinationOptions } from './helpers.ts'
 import { callbackify } from './internal/callbackify.js'
 import { TypedClient } from './internal/client.ts'
 import { CopyConditions } from './internal/copy-conditions.ts'
 import {
-  calculateEvenSplits,
   getScope,
   isBoolean,
   isFunction,
@@ -39,10 +35,7 @@ import {
   isValidObjectName,
   isValidPrefix,
   makeDateLong,
-  PART_CONSTRAINTS,
-  partsRequired,
   pipesetup,
-  sanitizeETag,
   uriEscape,
 } from './internal/helper.ts'
 import { PostPolicy } from './internal/post-policy.ts'
@@ -623,225 +616,6 @@ export class Client extends TypedClient {
 
     return listener
   }
-
-  /**
-   * Internal method to upload a part during compose object.
-   * @param partConfig __object__ contains the following.
-   *    bucketName __string__
-   *    objectName __string__
-   *    uploadID __string__
-   *    partNumber __number__
-   *    headers __object__
-   * @param cb called with null incase of error.
-   */
-  uploadPartCopy(partConfig, cb) {
-    const { bucketName, objectName, uploadID, partNumber, headers } = partConfig
-
-    const method = 'PUT'
-    let query = `uploadId=${uploadID}&partNumber=${partNumber}`
-    const requestOptions = { method, bucketName, objectName: objectName, query, headers }
-    return this.makeRequest(requestOptions, '', [200], '', true, (e, response) => {
-      let partCopyResult = Buffer.from('')
-      if (e) {
-        return cb(e)
-      }
-      pipesetup(response, transformers.uploadPartTransformer())
-        .on('data', (data) => {
-          partCopyResult = data
-        })
-        .on('error', cb)
-        .on('end', () => {
-          let uploadPartCopyRes = {
-            etag: sanitizeETag(partCopyResult.ETag),
-            key: objectName,
-            part: partNumber,
-          }
-
-          cb(null, uploadPartCopyRes)
-        })
-    })
-  }
-
-  composeObject(destObjConfig = {}, sourceObjList = [], cb) {
-    const me = this // many async flows. so store the ref.
-    const sourceFilesLength = sourceObjList.length
-
-    if (!Array.isArray(sourceObjList)) {
-      throw new errors.InvalidArgumentError('sourceConfig should an array of CopySourceOptions ')
-    }
-    if (!(destObjConfig instanceof CopyDestinationOptions)) {
-      throw new errors.InvalidArgumentError('destConfig should of type CopyDestinationOptions ')
-    }
-
-    if (sourceFilesLength < 1 || sourceFilesLength > PART_CONSTRAINTS.MAX_PARTS_COUNT) {
-      throw new errors.InvalidArgumentError(
-        `"There must be as least one and up to ${PART_CONSTRAINTS.MAX_PARTS_COUNT} source objects.`,
-      )
-    }
-
-    if (!isFunction(cb)) {
-      throw new TypeError('callback should be of type "function"')
-    }
-
-    for (let i = 0; i < sourceFilesLength; i++) {
-      if (!sourceObjList[i].validate()) {
-        return false
-      }
-    }
-
-    if (!destObjConfig.validate()) {
-      return false
-    }
-
-    const getStatOptions = (srcConfig) => {
-      let statOpts = {}
-      if (!_.isEmpty(srcConfig.VersionID)) {
-        statOpts = {
-          versionId: srcConfig.VersionID,
-        }
-      }
-      return statOpts
-    }
-    const srcObjectSizes = []
-    let totalSize = 0
-    let totalParts = 0
-
-    const sourceObjStats = sourceObjList.map((srcItem) =>
-      me.statObject(srcItem.Bucket, srcItem.Object, getStatOptions(srcItem)),
-    )
-
-    return Promise.all(sourceObjStats)
-      .then((srcObjectInfos) => {
-        const validatedStats = srcObjectInfos.map((resItemStat, index) => {
-          const srcConfig = sourceObjList[index]
-
-          let srcCopySize = resItemStat.size
-          // Check if a segment is specified, and if so, is the
-          // segment within object bounds?
-          if (srcConfig.MatchRange) {
-            // Since range is specified,
-            //    0 <= src.srcStart <= src.srcEnd
-            // so only invalid case to check is:
-            const srcStart = srcConfig.Start
-            const srcEnd = srcConfig.End
-            if (srcEnd >= srcCopySize || srcStart < 0) {
-              throw new errors.InvalidArgumentError(
-                `CopySrcOptions ${index} has invalid segment-to-copy [${srcStart}, ${srcEnd}] (size is ${srcCopySize})`,
-              )
-            }
-            srcCopySize = srcEnd - srcStart + 1
-          }
-
-          // Only the last source may be less than `absMinPartSize`
-          if (srcCopySize < PART_CONSTRAINTS.ABS_MIN_PART_SIZE && index < sourceFilesLength - 1) {
-            throw new errors.InvalidArgumentError(
-              `CopySrcOptions ${index} is too small (${srcCopySize}) and it is not the last part.`,
-            )
-          }
-
-          // Is data to copy too large?
-          totalSize += srcCopySize
-          if (totalSize > PART_CONSTRAINTS.MAX_MULTIPART_PUT_OBJECT_SIZE) {
-            throw new errors.InvalidArgumentError(`Cannot compose an object of size ${totalSize} (> 5TiB)`)
-          }
-
-          // record source size
-          srcObjectSizes[index] = srcCopySize
-
-          // calculate parts needed for current source
-          totalParts += partsRequired(srcCopySize)
-          // Do we need more parts than we are allowed?
-          if (totalParts > PART_CONSTRAINTS.MAX_PARTS_COUNT) {
-            throw new errors.InvalidArgumentError(
-              `Your proposed compose object requires more than ${PART_CONSTRAINTS.MAX_PARTS_COUNT} parts`,
-            )
-          }
-
-          return resItemStat
-        })
-
-        if ((totalParts === 1 && totalSize <= PART_CONSTRAINTS.MAX_PART_SIZE) || totalSize === 0) {
-          return this.copyObject(sourceObjList[0], destObjConfig, cb) // use copyObjectV2
-        }
-
-        // preserve etag to avoid modification of object while copying.
-        for (let i = 0; i < sourceFilesLength; i++) {
-          sourceObjList[i].MatchETag = validatedStats[i].etag
-        }
-
-        const splitPartSizeList = validatedStats.map((resItemStat, idx) => {
-          const calSize = calculateEvenSplits(srcObjectSizes[idx], sourceObjList[idx])
-          return calSize
-        })
-
-        function getUploadPartConfigList(uploadId) {
-          const uploadPartConfigList = []
-
-          splitPartSizeList.forEach((splitSize, splitIndex) => {
-            const { startIndex: startIdx, endIndex: endIdx, objInfo: objConfig } = splitSize
-
-            let partIndex = splitIndex + 1 // part index starts from 1.
-            const totalUploads = Array.from(startIdx)
-
-            const headers = sourceObjList[splitIndex].getHeaders()
-
-            totalUploads.forEach((splitStart, upldCtrIdx) => {
-              let splitEnd = endIdx[upldCtrIdx]
-
-              const sourceObj = `${objConfig.Bucket}/${objConfig.Object}`
-              headers['x-amz-copy-source'] = `${sourceObj}`
-              headers['x-amz-copy-source-range'] = `bytes=${splitStart}-${splitEnd}`
-
-              const uploadPartConfig = {
-                bucketName: destObjConfig.Bucket,
-                objectName: destObjConfig.Object,
-                uploadID: uploadId,
-                partNumber: partIndex,
-                headers: headers,
-                sourceObj: sourceObj,
-              }
-
-              uploadPartConfigList.push(uploadPartConfig)
-            })
-          })
-
-          return uploadPartConfigList
-        }
-
-        const performUploadParts = (uploadId) => {
-          const uploadList = getUploadPartConfigList(uploadId)
-
-          async.map(uploadList, me.uploadPartCopy.bind(me), (err, res) => {
-            if (err) {
-              this.abortMultipartUpload(destObjConfig.Bucket, destObjConfig.Object, uploadId).then(
-                () => cb(),
-                (err) => cb(err),
-              )
-              return
-            }
-            const partsDone = res.map((partCopy) => ({ etag: partCopy.etag, part: partCopy.part }))
-            return me.completeMultipartUpload(destObjConfig.Bucket, destObjConfig.Object, uploadId, partsDone).then(
-              (result) => cb(null, result),
-              (err) => cb(err),
-            )
-          })
-        }
-
-        const newUploadHeaders = destObjConfig.getHeaders()
-
-        me.initiateNewMultipartUpload(destObjConfig.Bucket, destObjConfig.Object, newUploadHeaders).then(
-          (uploadId) => {
-            performUploadParts(uploadId)
-          },
-          (err) => {
-            cb(err, null)
-          },
-        )
-      })
-      .catch((error) => {
-        cb(error, null)
-      })
-  }
 }
 
 Client.prototype.presignedUrl = promisify(Client.prototype.presignedUrl)
@@ -851,7 +625,6 @@ Client.prototype.presignedPostPolicy = promisify(Client.prototype.presignedPostP
 Client.prototype.getBucketNotification = promisify(Client.prototype.getBucketNotification)
 Client.prototype.setBucketNotification = promisify(Client.prototype.setBucketNotification)
 Client.prototype.removeAllBucketNotification = promisify(Client.prototype.removeAllBucketNotification)
-Client.prototype.composeObject = promisify(Client.prototype.composeObject)
 
 // refactored API use promise internally
 Client.prototype.makeBucket = callbackify(Client.prototype.makeBucket)
@@ -896,3 +669,4 @@ Client.prototype.getObjectRetention = callbackify(Client.prototype.getObjectRete
 Client.prototype.removeObjects = callbackify(Client.prototype.removeObjects)
 Client.prototype.removeIncompleteUpload = callbackify(Client.prototype.removeIncompleteUpload)
 Client.prototype.copyObject = callbackify(Client.prototype.copyObject)
+Client.prototype.composeObject = callbackify(Client.prototype.composeObject)
